requirements.txt
adbc_driver_manager==1.5.0
adbc_driver_postgresql==1.5.0
adbc_driver_sqlite==1.5.0
aiodns==3.2.0
anthropic==0.49.0
anywidget==0.9.18
AppKit==0.2.8
argcomplete==3.6.1
asgiref==3.8.1
async_timeout==5.0.1
atheris==2.3.0
awscrt==0.25.4
beautifulsoup4==4.13.3
bokeh==3.7.1
boto3==1.37.20
brotli==1.1.0
brotlicffi==1.1.0.0
checks==0.2
ConfigParser==7.2.0
contextlib2==21.6.0
cppyy==3.5.0
cryptography==44.0.2
ctypes_snappy==1.03
cudf==0.6.1.post1
cycler==0.12.1
Cython==3.0.12
dask_expr==2.0.0
defusedxml==0.7.1
distributed==2025.3.0
dl==0.1.0
docutils==0.21.2
duckdb==1.2.1
exceptiongroup==1.2.2
fastparquet==2024.11.0
filelock==3.18.0
fontTools==4.56.0
Foundation==0.1.0a0.dev1
fqdn==1.5.1
gitdb_speedups==0.1.0
gunicorn==23.0.0
h2==4.2.0
HTMLParser==0.0.2
hypothesis==6.130.4
ibis==3.3.0
importlib_resources==6.5.2
ini2toml==0.15
ipython==8.12.3
ipywidgets==8.1.5
isal==1.7.2
isoduration==20.11.0
jnius==1.1.0
JPype1==1.5.2
jsonpointer==3.0.0
kerchunk==0.2.8
keyring==25.6.0
langdetect==1.0.9
lxml==5.3.1
lz4==4.4.3
lzmaffi==0.3.0
marisa_trie==1.2.1
mask==1.0.0
matplotlib==3.10.1
mistralai==1.6.0
mock==5.2.0
modin==0.32.0
mtrand==0.1
mypy==1.15.0
numexpr==2.10.2
odfpy==1.4.1
olefile==0.47
openai==1.68.2
openpyxl==3.1.5
paramiko==3.5.1
pexpect==4.9.0
pickle5==0.0.12
pkg1==0.0.3
pkgutil_resolve_name==1.3.10
polars==1.26.0
psutil==7.0.0
pycares==4.5.0
pycurl==7.45.6
pygit2==1.17.0
PyInstaller==6.12.0
pymupdf_fonts==1.0.5
pyodide==0.0.2
pyOpenSSL==25.0.0
pyperf==2.9.0
PyQt4==4.11.4
PyQt5==5.15.11
PyQt6==6.8.1
PySide6==6.8.2.1
pyspark==3.5.5
python_calamine==0.3.1
pyxlsb==1.0.10
PyYAML==6.0.2
qtpy==2.4.3
railroad==0.5.0
redis==5.2.1
requests_kerberos==0.15.0
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rfc3987==1.3.8
scikit_learn==1.6.1
scipy==1.15.2
scipy_doctest==1.6
sets==0.3.2
setuptools_scm==8.2.0
simplejson==3.20.1
skimage==0.0
smbprotocol==1.15.0
sniffio==1.3.1
snowflake==1.1.0
Sphinx==8.2.3
SQLAlchemy==2.0.39
statsmodels==0.14.4
striprtf==0.0.28
sympy==1.13.3
tables==3.10.2
thread==2.0.5
threadpoolctl==3.6.0
tomli_w==1.2.0
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trove_classifiers==2025.3.19.19
ujson==5.10.0
uri_template==1.3.0
urllib3_secure_extra==0.1.0
uvloop==0.21.0
vegafusion==2.0.2
vincent==0.4.4
webcolors==24.11.1
xarray==2025.3.0
xlrd==2.0.1
xlsxwriter==3.2.2
xmlrpclib==1.0.1
xx==3.3.2
zope==5.13
zstandard==0.23.0
